Move getPathsDirectory functionality
