+++ tools/cli/helpers/util.js
@@ -174,6 +174,17 @@ function moveToModules(directory) {
   return shell.pwd().stdout;
 }
 
+function getPathsDirectory(route) {
+  const listDirPaths = [];
+  const elements = fs.readdirSync(route);
+  elements.forEach(element => {
+    if (!fs.statSync(`${route}/${element}`).isFile()) {
+      listDirPaths.push(`${route}/${element}`);
+    }
+  });
+  return listDirPaths;
+}
+
 module.exports = {
   getModulePackageName,
   getTemplatesPath,
@@ -188,5 +199,6 @@ module.exports = {
   addSymlink,
   removeSymlink,
   runPrettier,
-  moveToModules
+  moveToModules,
+  getPathsDirectory
 };
